+++ packages/iov-core/types/multichainsigner.d.ts
@@ -1,5 +1,5 @@
 import { Address, BcpConnection, BcpTransactionResponse, ChainConnector, Nonce, UnsignedTransaction } from "@iov/bcp-types";
-import { KeyringEntryId, UserProfile } from "@iov/keycontrol";
+import { UserProfile, WalletId } from "@iov/keycontrol";
 import { ChainId, PublicKeyBundle } from "@iov/tendermint-types";
 export declare class MultiChainSigner {
     readonly profile: UserProfile;
@@ -17,7 +17,7 @@ export declare class MultiChainSigner {
     }>;
     keyToAddress(chainId: ChainId, key: PublicKeyBundle): Address;
     getNonce(chainId: ChainId, addr: Address): Promise<Nonce>;
-    signAndCommit(tx: UnsignedTransaction, keyring: KeyringEntryId): Promise<BcpTransactionResponse>;
+    signAndCommit(tx: UnsignedTransaction, walletId: WalletId): Promise<BcpTransactionResponse>;
     /**
      * Throws for unknown chain ID
      */