+++ packages/iov-keycontrol/src/index.ts
@@ -2,12 +2,12 @@ export { HdPaths } from "./hdpaths";
 export { Ed25519HdWallet, Ed25519KeyringEntry, Secp256k1HdWallet } from "./keyring-entries";
 export {
   Keyring,
-  KeyringEntry,
-  KeyringEntryId,
-  KeyringEntryImplementationIdString,
-  KeyringEntrySerializationString,
   LocalIdentity,
   LocalIdentityId,
   PublicIdentity,
+  Wallet,
+  WalletId,
+  WalletImplementationIdString,
+  WalletSerializationString,
 } from "./keyring";
 export { UserProfile, WalletInfo } from "./userprofile";