+++ packages/iov-keycontrol/types/index.d.ts
@@ -1,4 +1,4 @@
 export { HdPaths } from "./hdpaths";
 export { Ed25519HdWallet, Ed25519KeyringEntry, Secp256k1HdWallet } from "./keyring-entries";
-export { Keyring, KeyringEntry, KeyringEntryId, KeyringEntryImplementationIdString, KeyringEntrySerializationString, LocalIdentity, LocalIdentityId, PublicIdentity, } from "./keyring";
+export { Keyring, LocalIdentity, LocalIdentityId, PublicIdentity, Wallet, WalletId, WalletImplementationIdString, WalletSerializationString, } from "./keyring";
 export { UserProfile, WalletInfo } from "./userprofile";